requirements.txt
absl-py==2.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cmdstanpy==1.2.4
contourpy==1.3.0
cycler==0.12.1
Deprecated==1.2.14
executing==2.1.0
fastapi==0.115.4
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
graphql-core==3.2.5
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
holidays==0.60
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
libcst==1.5.0
logfire==2.1.2
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-instrumentation-sqlalchemy==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
opt_einsum==3.4.0
optree==0.13.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prophet==1.1.6
protobuf==5.28.3
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stanio==0.5.1
starlette==0.41.2
strawberry-graphql==0.248.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.1.3
wheel==0.45.0
wrapt==1.16.0
zipp==3.21.0
